% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm_strlength.R
\name{choose_strlen}
\alias{choose_strlen}
\title{choose string length}
\usage{
choose_strlen(mdb, len)
}
\arguments{
\item{mdb}{dataframe. output from \code{confirm_strlen}}

\item{len}{integer vector.}
}
\value{
dataframe with original columns, filtered to the specific string length
}
\description{
a helper function for \code{confirm_strlen} that filters the output dataframe on a specific string length for that column
}
