% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaner.R
\name{cleaner}
\alias{cleaner}
\title{Cleans Incoming Data from Excel}
\usage{
cleaner(d)
}
\arguments{
\item{d}{data to be cleaned}
}
\description{
Cleans incoming data a lot of invisible NAs can get imputed in Excel that need to be removed
}
\examples{
\dontrun{
ll <- read_csv("~/Desktop/file.csv")
ll <- cleaner(ll)
}
}

