% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm_overlap_helpers.R
\name{co_find_in_both}
\alias{co_find_in_both}
\title{confirm overlap - find entries in both dfs}
\usage{
co_find_in_both(mdb)
}
\arguments{
\item{mdb}{dataframe output from confirm_overlap}
}
\value{
dataframe
}
\description{
returns a tibble whose column are entries shared between both columns
}
