% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm_overlap_helpers.R
\name{co_find_only_in_2}
\alias{co_find_only_in_2}
\title{confirm overlap - find entries only in df 2}
\usage{
co_find_only_in_2(mdb)
}
\arguments{
\item{mdb}{dataframe output from confirm_overlap}
}
\value{
dataframe
}
\description{
returns a tibble whose column are entries unique to the column in the second df
}
