% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_compare.R
\name{compare}
\alias{compare}
\title{Main Helper Function for vc_compare}
\usage{
compare(xx, yy, fun)
}
\arguments{
\item{xx}{character vector to be coerced}

\item{fun}{logical operator}

\item{yx}{character vector to be coerced and compared}
}
\description{
Helper Function for vc_compare
}

