% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm_mapping.R
\name{confirm_mapping}
\alias{confirm_mapping}
\title{Confirm structural mapping between 2 columns}
\usage{
confirm_mapping(.data, col1, col2, view = T)
}
\arguments{
\item{.data}{a data frame}

\item{col1}{column 1}

\item{col2}{column 2}

\item{view}{View results?}
}
\value{
A view of mappings. Also returns the view as a data frame invisibly.
}
\description{
The mapping between elements of 2 columns can have 4 different relationships: one - one, one - many, many - one, many - many.
This function returns a view of the mappings by row, and prints a summary to the console.
}
\examples{
iris \%>\% confirm_mapping(Species, Sepal.Width, view = FALSE)
}
