% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm_overlap.R
\name{confirm_overlap_internal}
\alias{confirm_overlap_internal}
\title{Confirm Overlap internal}
\usage{
confirm_overlap_internal(vec1, vec2)
}
\arguments{
\item{vec1}{vector 1}

\item{vec2}{vector 2}
}
\value{
1 row tibble
}
\description{
A venn style summary of the overlap in unique values of 2 vectors
}
\examples{
confirm_overlap(iris$Sepal.Width, iris$Sepal.Length)
}
\keyword{internal}
