% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm_strlength.R
\name{confirm_strlen}
\alias{confirm_strlen}
\title{confirm string length}
\usage{
confirm_strlen(mdb, col)
}
\arguments{
\item{mdb}{dataframe}

\item{col}{unquoted column}
}
\value{
prints a summary and returns a dataframe invisibly
}
\description{
returns a count table of string lengths for a character column.
}
\examples{
iris \%>\% confirm_strlen(Species)
}
