% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_core_structure.R
\name{convert_to_core_structure}
\alias{convert_to_core_structure}
\title{Convert Directory of Predictives Files to Core Folder Structure}
\usage{
convert_to_core_structure(
  path,
  institution = "Institution",
  first.import = FALSE,
  exdir = file.path(tempdir(), institution),
  ...
)
}
\arguments{
\item{path}{A path to the directory where the predictives files reside.}

\item{institution}{A string representing the institution's name.}

\item{first.import}{logical. If `TRUE` then StudentImportPrior.csv file is not
required, otherwise a hard stop occurs if this file is missing.}

\item{exdir}{The directory to extract files to. NOTE that this folder will be
deleted and then recreated.  You will lose any data cotnained within.
Defaults to creating a folder named the same as `institution` in the system's
temp directory.}

\item{\ldots}{ignored.}
}
\value{
Returns the path to exdir for easy piping.
}
\description{
Convert a single directory of .csv files to a directory of core data folders
and files.
}
\examples{
\dontrun{
temp_loc <- as_core_structure(
    path = 'data/tyler_fake',
    institution = 'HomeTown',
    first.import = TRUE
)

cl::opener(temp_loc)
}
}
