% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{data_mode}
\alias{data_mode}
\title{data_mode}
\usage{
data_mode(x, prop = TRUE)
}
\arguments{
\item{x}{vector}

\item{prop}{show frequency as ratio? default T}
}
\value{
named double of length 1
}
\description{
data_mode
}
\keyword{internal}
