% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataValidation-package.R
\docType{package}
\name{dataValidation-package}
\alias{dataValidation}
\alias{dataValidation-package}
\title{dataValidation: Validation For Data Frames}
\description{
Functions for validating the structure and properties of data frames.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://harrison4192.github.io/dataValidation}
  \item \url{https://github.com/Harrison4192/dataValidation}
  \item Report bugs at \url{https://github.com/Harrison4192/dataValidation/issues}
}

}
\author{
\strong{Maintainer}: Harrison Tietze \email{Harrison4192@gmail.com}

}
\keyword{internal}
