% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_cardinality.R
\name{determine_cardinality}
\alias{determine_cardinality}
\title{Determine pairwise structural mappings}
\usage{
determine_cardinality(df, ...)
}
\arguments{
\item{df}{a data frame}

\item{...}{columns or a tidyselect specification}
}
\value{
description of mappings
}
\description{
Determine pairwise structural mappings
}
\examples{
iris \%>\% determine_cardinality(everything())
}
