% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose}
\alias{diagnose}
\title{diagnose}
\usage{
diagnose(df, ...)
}
\arguments{
\item{df}{dataframe}

\item{...}{tidyselect}
}
\value{
dataframe summary
}
\description{
Pipe in a dataframe to return a diagnosis of its missing and unique values for each columns.
Default behavior is to diagnose all columns, but a subset can be specified in the dots with tidyselect.
}
\details{
this function is inspired by the excellent \href{https://choonghyunryu.github.io/dlookr/}{dlookr} package. It takes a dataframe and returns
a summary of unique and missing values of the columns.
}
\examples{
iris \%>\% diagnose()
}
