% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose_category}
\alias{diagnose_category}
\title{diagnose category}
\usage{
diagnose_category(.data, ..., max_distinct = 5)
}
\arguments{
\item{.data}{dataframe}

\item{max_distinct}{integer}
}
\value{
dataframe
}
\description{
diagnose category
}
