% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose_numeric}
\alias{diagnose_numeric}
\title{diagnose_numeric}
\usage{
diagnose_numeric(.data, ...)
}
\arguments{
\item{.data}{dataframe}

\item{...}{tidyselect}
}
\value{
dataframe
}
\description{
diagnose_numeric
}
