% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_useless_cols.R
\name{drop_useless_cols}
\alias{drop_useless_cols}
\title{Drop useless columns}
\usage{
drop_useless_cols(.data)
}
\arguments{
\item{.data}{a data frame}
}
\value{
invisibly returns data frame
}
\description{
Removes columns that are all NA or contain only 1 value.
}
\examples{
iris \%>\% drop_useless_cols
}
\keyword{internal}
