% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Reads in Imports Dictionary}
\usage{
import(file, method = "excel", ...)
}
\arguments{
\item{file}{path to file}

\item{method}{choice between reading dictionary from local or GoogleSheets source}

\item{\ldots}{extra params}
}
\value{
list containing table, file, and column-level dictionaries
}
\description{
Reads in Imports Dictionary
}
\examples{
\dontrun{
ll <- import("~/Desktop/Dictionary.xlsx")
}
}

