% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_map.R
\name{import_map}
\alias{import_map}
\title{Convenience Function to Imports and Map Dictionary}
\usage{
import_map(file, method = "excel", ...)
}
\arguments{
\item{file}{path to file}

\item{method}{choice between reading dictionary from local or GoogleSheets source}

\item{\ldots}{extra params}
}
\description{
Convenience Function to Imports and Map Dictionary: wraps up import() and mapper()
}
\examples{
\dontrun{
ll <- import_map("~/Desktop/Dictionary.xlsx")
}
}

