% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_distinct.R
\name{make_distincts}
\alias{make_distincts}
\title{Make Distinct}
\usage{
make_distincts(df, ...)
}
\arguments{
\item{df}{a df}

\item{...}{cols}
}
\value{
a list of name lists
}
\description{
Make Distinct
}
\keyword{internal}
