% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapper.R
\name{mapper}
\alias{mapper}
\title{Creates Dictionary Mapping}
\usage{
mapper(dict_list)
}
\arguments{
\item{dict_list}{dictionary list produced by import()}
}
\description{
Creates Dictionary Mapping as List of Lists of Functions
}
\examples{
\dontrun{
ll <- import("~/Desktop/Dictionary.xlsx")
ll <- mapper(ll)
}
}

