% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_distinct.R
\name{names_list}
\alias{names_list}
\title{Names List}
\usage{
names_list(df, len)
}
\arguments{
\item{df}{a df}

\item{len}{how many elements in combination}
}
\value{
a list of name combinations
}
\description{
Names List
}
\keyword{internal}
