% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_compare.R
\name{not-equal}
\alias{not-equal}
\alias{!=@}
\title{Not Equal Ignoring Case}
\usage{
"!=@"(x, y)
}
\arguments{
\item{x}{number one}

\item{y}{number two}
}
\description{
Helper Function for vc_compare to implement not equal with case ignored
}
