% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/printing_functions.R
\name{print.all_good}
\alias{print.all_good}
\title{Prints a all_good Object}
\usage{
\method{print}{all_good}(x, ...)
}
\arguments{
\item{x}{An \code{all_good} object.}

\item{\ldots}{ignored.}
}
\description{
Prints a all_good object
}

