% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/printing_functions.R
\name{print.invalid_report}
\alias{print.invalid_report}
\title{Prints a invalid_report Object}
\usage{
\method{print}{invalid_report}(x, ...)
}
\arguments{
\item{x}{An \code{invalid_report} object.}

\item{\ldots}{ignored.}
}
\description{
Prints a invalid_report object
}

