% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/printing_functions.R
\name{print.valid_report}
\alias{print.valid_report}
\title{Prints a valid_report Object}
\usage{
\method{print}{valid_report}(x, ...)
}
\arguments{
\item{x}{A \code{valid_report} object.}

\item{\ldots}{ignored.}
}
\description{
Prints a valid_report object
}

