% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valiData.R
\name{print.valiData}
\alias{print.valiData}
\title{Prints a valiData Object}
\usage{
\method{print}{valiData}(x, as.report = FALSE, delete = TRUE, ...)
}
\arguments{
\item{x}{A valiData object.}

\item{as.report}{logical.  If \code{TRUE} a report will be generated in the
directory indicated by \code{x$path}.}

\item{delete}{logical.  If \code{TRUE} and \code{as.report = TRUE} any prior
instances of `report will be deleted.}

\item{\ldots}{ignored.}
}
\description{
Prints a valiData object
}

