% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vd_dir.R
\name{print.vd_dir}
\alias{print.vd_dir}
\title{Prints a vd_dir Object}
\usage{
\method{print}{vd_dir}(x, ...)
}
\arguments{
\item{x}{A vd_dir object.}

\item{\ldots}{ignored.}
}
\description{
Prints a vd_dir object
}

