% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vd_empty_subfolders.R
\name{print.vd_empty_subfolders}
\alias{print.vd_empty_subfolders}
\title{Prints a vd_empty_subfolders Object}
\usage{
\method{print}{vd_empty_subfolders}(x, ...)
}
\arguments{
\item{x}{A vd_empty_subfolders object.}

\item{\ldots}{ignored.}
}
\description{
Prints a vd_empty_subfolders object
}

