% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vd_multiple_csv.R
\name{print.vd_multiple_csv}
\alias{print.vd_multiple_csv}
\title{Prints a vd_multiple_csv Object}
\usage{
\method{print}{vd_multiple_csv}(x, ...)
}
\arguments{
\item{x}{A vd_multiple_csv object.}

\item{\ldots}{ignored.}
}
\description{
Prints a vd_multiple_csv object
}
