% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vd_non_empty.R
\name{print.vd_non_empty}
\alias{print.vd_non_empty}
\title{Prints a vd_non_empty Object}
\usage{
\method{print}{vd_non_empty}(x, ...)
}
\arguments{
\item{x}{A vd_non_empty object.}

\item{\ldots}{ignored.}
}
\description{
Prints a vd_non_empty object
}

