% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vd_unmapped_csv.R
\name{print.vd_unmapped_csv}
\alias{print.vd_unmapped_csv}
\title{Prints a vd_unmapped_csv Object}
\usage{
\method{print}{vd_unmapped_csv}(x, ...)
}
\arguments{
\item{x}{A vd_unmapped_csv object.}

\item{\ldots}{ignored.}
}
\description{
Prints a vd_unmapped_csv object
}

