% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vf_csv_broken.R
\name{print.vf_csv_broken}
\alias{print.vf_csv_broken}
\title{Prints a vf_csv_broken Object}
\usage{
\method{print}{vf_csv_broken}(x, ...)
}
\arguments{
\item{x}{A vf_csv_broken object.}

\item{\ldots}{ignored.}
}
\description{
Prints a vf_csv_broken object
}

