% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vf_embedded_nul.R
\name{print.vf_embedded_nul}
\alias{print.vf_embedded_nul}
\title{Prints a vf_embedded_nul Object}
\usage{
\method{print}{vf_embedded_nul}(x, ...)
}
\arguments{
\item{x}{A vf_embedded_nul object.}

\item{\ldots}{ignored.}
}
\description{
Prints a vf_embedded_nul object
}
