% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vf_file_type.R
\name{print.vf_file_type}
\alias{print.vf_file_type}
\title{Prints a vf_file_type Object}
\usage{
\method{print}{vf_file_type}(x, ...)
}
\arguments{
\item{x}{A vf_file_type object.}

\item{\ldots}{ignored.}
}
\description{
Prints a vf_file_type object
}

