% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vf_non_empty.R
\name{print.vf_non_empty}
\alias{print.vf_non_empty}
\title{Prints a vf_non_empty Object}
\usage{
\method{print}{vf_non_empty}(x, ...)
}
\arguments{
\item{x}{A vf_non_empty object.}

\item{\ldots}{ignored.}
}
\description{
Prints a vf_non_empty object
}

