% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_column_names.R
\name{print.vt_column_names}
\alias{print.vt_column_names}
\title{Prints a vt_column_names Object}
\usage{
\method{print}{vt_column_names}(x, ...)
}
\arguments{
\item{x}{A vt_column_names object.}

\item{\ldots}{ignored.}
}
\description{
Prints a vt_column_names object
}

