% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_column_order.R
\name{print.vt_column_order}
\alias{print.vt_column_order}
\title{Prints a vt_column_order Object}
\usage{
\method{print}{vt_column_order}(x, ...)
}
\arguments{
\item{x}{A vt_column_order object.}

\item{\ldots}{ignored.}
}
\description{
Prints a vt_column_order object
}

