% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_duplicated_rows.R
\name{print.vt_duplicated_rows}
\alias{print.vt_duplicated_rows}
\title{Prints a vt_duplicated_rows Object}
\usage{
\method{print}{vt_duplicated_rows}(x, ...)
}
\arguments{
\item{x}{A vt_duplicated_rows object.}

\item{\ldots}{ignored.}
}
\description{
Prints a vt_duplicated_rows object
}

