% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_enough_columns.R
\name{print.vt_enough_columns}
\alias{print.vt_enough_columns}
\title{Prints a vt_enough_columns Object}
\usage{
\method{print}{vt_enough_columns}(x, ...)
}
\arguments{
\item{x}{A vt_enough_columns object.}

\item{\ldots}{ignored.}
}
\description{
Prints a vt_enough_columns object
}

