% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_header.R
\name{print.vt_header}
\alias{print.vt_header}
\title{Prints a vt_header Object}
\usage{
\method{print}{vt_header}(x, ...)
}
\arguments{
\item{x}{A vt_header object.}

\item{\ldots}{ignored.}
}
\description{
Prints a vt_header object
}

