% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_na_rows.R
\name{print.vt_na_rows}
\alias{print.vt_na_rows}
\title{Prints a vt_na_rows Object}
\usage{
\method{print}{vt_na_rows}(x, ...)
}
\arguments{
\item{x}{A vt_na_rows object.}

\item{\ldots}{ignored.}
}
\description{
Prints a vt_na_rows object
}

