% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_non_ASCII.R
\name{print.vt_non_ASCII}
\alias{print.vt_non_ASCII}
\title{Prints a vt_non_ASCII  Object}
\usage{
\method{print}{vt_non_ASCII}(x, ...)
}
\arguments{
\item{x}{A vt_non_ASCII  object.}

\item{\ldots}{ignored.}
}
\description{
Prints a vt_non_ASCII  object
}

