% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_non_empty.R
\name{print.vt_non_empty}
\alias{print.vt_non_empty}
\title{Prints a vt_non_empty  Object}
\usage{
\method{print}{vt_non_empty}(x, ...)
}
\arguments{
\item{x}{A vt_non_empty  object.}

\item{\ldots}{ignored.}
}
\description{
Prints a vt_non_empty  object
}

