% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_required_columns.R
\name{print.vt_required_columns}
\alias{print.vt_required_columns}
\title{Prints a vt_required_columns  Object}
\usage{
\method{print}{vt_required_columns}(x, ...)
}
\arguments{
\item{x}{A vt_required_columns  object.}

\item{\ldots}{ignored.}
}
\description{
Prints a vt_required_columns  object
}

