% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_required_headers.R
\name{print.vt_required_headers}
\alias{print.vt_required_headers}
\title{Prints a vt_required_headers Object}
\usage{
\method{print}{vt_required_headers}(x, ...)
}
\arguments{
\item{x}{A vt_required_headers object.}

\item{\ldots}{ignored.}
}
\description{
Prints a vt_required_headers object
}
