% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_spaced_colnames.R
\name{print.vt_spaced_colnames}
\alias{print.vt_spaced_colnames}
\title{Prints a vt_spaced_colnames  Object}
\usage{
\method{print}{vt_spaced_colnames}(x, ...)
}
\arguments{
\item{x}{A vt_spaced_colnames  object.}

\item{\ldots}{ignored.}
}
\description{
Prints a vt_spaced_colnames  object
}

