% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_column_map_dir.R
\name{read_column_map_dir}
\alias{read_column_map_dir}
\title{Read in Debian Control File}
\usage{
read_column_map_dir(path, names = NULL)
}
\arguments{
\item{path}{Path to directory containing multiple Debian Control files.}

\item{names}{Option vector of names to add to the list output.  Defualt uses
the name of the Debian Control files sans file extenson.}
}
\value{
Returns a list of named column tests to be run in \code{valiData}.
}
\description{
Read in Debian Control file column mapping.
}

