% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_all_is_well.R
\name{report_all_is_well}
\alias{report_all_is_well}
\title{Generate Random Cowsay Praise}
\usage{
report_all_is_well()
}
\description{
Generate random praise from a cow.
}
\examples{
report_all_is_well()
}

