% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data1}
\alias{sample_data1}
\title{Sample Data}
\format{
6 columns. 3 id and 3 values
\describe{
\item{ID_COL1}{4-5 distinct codes}
}
}
\usage{
sample_data1
}
\description{
Sample Data
}
\keyword{125}
\keyword{data}
\keyword{internal}
\keyword{rows}
\keyword{sample}
\keyword{with}
