% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_compare.R
\name{~=}
\alias{~=}
\title{Nearly Equal}
\usage{
"~="(x, y)
}
\arguments{
\item{x}{number one}

\item{y}{number two}
}
\description{
Helper Function for vc_compare to implement all.equal as new comparison operator
}

