% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/valiData.R
\name{valiData}
\alias{valiData}
\title{Validate a Directory of Subdirectories of CSV Files}
\usage{
valiData(path, core_data_map, column_map, delete = TRUE)
}
\arguments{
\item{path}{Path to a directory with drecories of .csv files.}

\item{core_data_map}{A mapping (\code{list} of \code{data.frame}s).  One
\code{data.frame} per .csv file with a \code{header} (header name) &
\code{required} (logical,; Is column required?).}

\item{column_map}{A list (one per file) of lists (one per column in each file)
of the desired tests to run on columns (see \code{read_column_map_dir}).}

\item{delete}{logical.  If \code{TRUE} the old report is deleted.}
}
\value{
Returns a report \file{~/`Reports/valiData_report.txt}.
}
\description{
Validate a directory of dubdirectories of .csv files.
}
\examples{
\dontrun{
valiData("C:/Users/trinker/Desktop/myfolder", mymap)
}
}
\seealso{
\code{\link[valiData]{validate_file}}
}

