% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_.R
\name{vc_}
\alias{vc_}
\title{Generic Column Validation}
\usage{
vc_(data, x, expr, augment_comment = NULL, ...)
}
\arguments{
\item{data}{A data frame.}

\item{x}{Column name from \code{data} (character string).}

\item{expr}{An string that can be evaluated elementwise as an expression to
test on \code{x} column (include column \code{x} as seen in Examples below).}

\item{augment_comment}{details to be added to message, if any}

\item{\dots}{ignored.}
}
\value{
Returns a \code{vc} classed list object.
}
\description{
Generic column validation
}
\examples{
vc_(mtcars, 'mpg', 'abs(scale(mpg)) < 2')
vc_(mtcars, 'mpg', 'abs(scale(mpg)) < 2',
    augment_comment = "...\\n    because they may be (outliers > 2 sd)")
vc_(CO2, 'Type', '!grepl("^M", Type)',
    augment_comment = " because they don't begin with `M`")
vc_(mtcars, 'drat', 'drat >= wt',
    augment_comment = " because `drat` is less than `wt`")
str(vc_(mtcars, 'mpg', 'abs(scale(mpg)) < 2'))
}

