% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_categories.R
\name{vc_categories}
\alias{vc_categories}
\title{Validates If Correct Categories Used}
\usage{
vc_categories(data, x, levels = "the levels", ...)
}
\arguments{
\item{data}{A data frame.}

\item{x}{Column name from \code{data} (character string).}

\item{levels}{Levels of the category.}

\item{\dots}{ignored.}
}
\value{
Returns a \code{vc} classed list object.
}
\description{
Validates If Correct Categories Used
}
\examples{
vc_categories(mtcars, 'cyl', c(6, 8))
str(vc_categories(mtcars, 'cyl', c(6, 8)))
}

