% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_cipcode.R
\name{vc_cipcode}
\alias{vc_cipcode}
\title{Validates and Reports If Valid CIP Code}
\usage{
vc_cipcode(x, colname_x = "the column")
}
\arguments{
\item{x}{character vector}

\item{colname_x}{vector's colname}
}
\description{
Validates and Reports If Valid CIP Code
}

