% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_column_apply.R
\name{vc_column_apply}
\alias{vc_column_apply}
\title{Apply Column Map to Dataframe}
\usage{
vc_column_apply(data, colmap)
}
\arguments{
\item{data}{A data frame to run column-wise tests on.}

\item{colmap}{A column map from \code{read_column_map_dir}.}
}
\description{
Apply a column map from \code{read_column_map_dir} to a
\code{\link[base]{data.frame}}.
}

