% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vc_credits.R
\name{vc_credits}
\alias{vc_credits}
\title{Validates and Reports If Formatted Like Credit}
\usage{
vc_credits(x, colnames = "the column")
}
\arguments{
\item{x}{character vector}

\item{colnames}{vector's colname}
}
\description{
Validates and Reports If Formatted Like Credit
}

