% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_email.R
\name{vc_email}
\alias{vc_email}
\title{Validates and Reports If Email}
\usage{
vc_email(x, colname_x = "the column")
}
\arguments{
\item{x}{character vector}

\item{colname_x}{vector's colname}
}
\description{
Validates and Reports If Email
}

