% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_greater_than.R
\name{vc_greater_than}
\alias{vc_greater_than}
\title{Validates and Reports If Greater Than}
\usage{
vc_greater_than(x, y, colname_x = "the X column",
  colname_y = "the Y column")
}
\arguments{
\item{x}{character vector to be coerced}

\item{y}{character vector to be coerced and compared}

\item{colname_x}{X vector's colname}

\item{colname_y}{Y vector's colname}
}
\description{
Validates and Reports If Greater Than
}

