% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_iso_datetime.R
\name{vc_iso_datetime}
\alias{vc_iso_datetime}
\title{Validates and Reports If ISO 8601}
\usage{
vc_iso_datetime(data, x, ...)
}
\arguments{
\item{data}{A data frame.}

\item{x}{Column name from \code{data} (character string).}

\item{\dots}{ignored.}
}
\description{
Validates and Reports If ISO 8601
}
\examples{
dat <- data.frame(
    start = c("2016-01-01T09", "R2D2-3CPO", sprintf("2016-04-\%sT09", 12:19), NA),
    end = c(NA, sprintf("2016-01-\%sT09", 11:20)),
    stringsAsFactors = FALSE
)

vc_iso_datetime(dat, 'start')
vc_iso_datetime(dat, 'end')
}

