% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vc_leq.R
\name{vc_leq}
\alias{vc_leq}
\title{Validates and Reports If Less Than Or Equal To}
\usage{
vc_leq(x, y, colname_x = "the X column", colname_y = "the Y column")
}
\arguments{
\item{x}{character vector to be coerced}

\item{y}{character vector to be coerced and compared}

\item{colname_x}{X vector's colname}

\item{colname_y}{Y vector's colname}
}
\description{
Validates and Reports If Less Than Or Equal To
}

