% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vc_less_than.R
\name{vc_less_than}
\alias{vc_less_than}
\title{Validates and Reports If Less Than}
\usage{
vc_less_than(x, y, colnames = "the X column", colname_y = "the Y column")
}
\arguments{
\item{x}{character vector to be coerced}

\item{y}{character vector to be coerced and compared}

\item{colnames}{X vector's colname}

\item{colname_y}{Y vector's colname}
}
\description{
Validates and Reports If Less Than
}

