% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_logical.R
\name{vc_logical}
\alias{vc_logical}
\title{Validates If Logical}
\usage{
vc_logical(data, x, ...)
}
\arguments{
\item{data}{A data frame.}

\item{x}{Column name from \code{data} (character string).}

\item{\dots}{ignored.}
}
\description{
Validates If Logical
}
\examples{
dat <- data.frame(
    check = c('T', 'TRUE', 'FALSE', 'F', 'false', 'true', NA, TRUE, FALSE, 'TURE'),
    stringsAsFactors = FALSE
 )
vc_logical(dat, 'check')
}

