% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_nchar.R
\name{vc_nchar}
\alias{vc_nchar}
\title{Validates Number of Characters}
\usage{
vc_nchar(data, x, nchar, ...)
}
\arguments{
\item{data}{A data frame.}

\item{x}{Column name from \code{data} (character string).}

\item{nchar}{The expected max number of characters.}

\item{\dots}{ignored.}
}
\description{
Validates Number of Characters
}
\examples{
dat <- data.frame(
    x = c('dog', 'cat', 'chicken', 'a', NA),
    stringsAsFactors = FALSE
 )
vc_nchar(dat, 'x', 3)
vc_nchar(dat, 'x', 10)
}

