% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_non_response.R
\name{vc_non_response}
\alias{vc_non_response}
\title{Checks For Non-Response Elements In Vector}
\usage{
vc_non_response(x, prop.acceptable = 0, missing = c("", "NULL", "NA", "N/A",
  "na", "n/a"), required = FALSE, file.name = NULL)
}
\arguments{
\item{x}{char vector}

\item{prop.acceptable}{proportion of allowable non-reponse}

\item{missing}{values non-response can take}

\item{required}{whether the vector in the table is required}

\item{file.name}{name of file containing the table of vectors}
}
\description{
Checks For Non-Response Elements In Vector
}
\examples{
set.seed(10)
x <- sample(c(NA, rep(LETTERS[1:10], 100)), 10000, TRUE)
vc_non_response(x)
}

