% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_null.R
\name{vc_null}
\alias{vc_null}
\title{Checks For Null Elements In Vector}
\usage{
vc_null(data, x, prop.acceptable = 0, missing = c("", "NULL"),
  required = FALSE, ...)
}
\arguments{
\item{data}{A data frame.}

\item{x}{Column name from \code{data} (character string).}

\item{prop.acceptable}{proportion of allowable missing.}

\item{missing}{values that missing can take.}

\item{required}{whether the vector in the table is required.}

\item{\dots}{ignored.}
}
\description{
Checks For Null Elements In Vector
}
\examples{
set.seed(10)
dat <- data.frame(x = sample(c("NULL", rep(LETTERS[1:10])), 10000, TRUE))
vc_null(dat, 'x')
vc_null(dat, 'x', prop.acceptable = .1)
}

