% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vc_numeric.R
\name{vc_numeric}
\alias{vc_numeric}
\title{Validates and Reports If Numeric}
\usage{
vc_numeric(x, colnames = "the column")
}
\arguments{
\item{x}{character vector}

\item{colnames}{vector's colname}
}
\description{
Validates and Reports If Numeric
}

