% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_regex.R
\name{vc_regex}
\alias{vc_regex}
\title{Validates and Reports If Formatted Like Regex Provided}
\usage{
vc_regex(data, x, regex, ...)
}
\arguments{
\item{data}{A data frame.}

\item{x}{Column name from \code{data} (character string).}

\item{regex}{the regular expression to be tested against.}

\item{\dots}{ignored.}
}
\description{
Validates and Reports If Formatted Like Regex Provided
}
\examples{
dat <- data.frame(
    vals = c('Greg', 'John', 'Tyrone', 'Susan', 'Kyra', 'Nells12', NA, ""),
    stringsAsFactors = FALSE
 )
vc_regex(dat, 'vals', '^[A-Z][^0-9]+$')
vc_regex(dat, 'vals', '^[A-Z].+$')
}

