% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_type.R
\name{vc_type}
\alias{vc_type}
\title{Switch Method Based On Type Of Data in Column}
\usage{
vc_type(data, x, type, ...)
}
\arguments{
\item{data}{A data frame.}

\item{x}{Column name from \code{data} (character string).}

\item{type}{:  date_utc, datetime, email, enumerated, integer, logical, numeric, zipcode, cipcode
guid,enumerated are handled outside by \code{vc_unique} and \code{vc_categories}}

\item{\dots}{ignored.}
}
\description{
Switch Method Based On Type Of Data in Column
}

