% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vc_utc_date.R
\name{vc_utc_date}
\alias{vc_utc_date}
\title{Validates and Reports If Date UTC}
\usage{
vc_utc_date(x, colnames = "the column")
}
\arguments{
\item{x}{character vector}

\item{colnames}{vector's colname}
}
\description{
Validates and Reports If Date UTC
}

