% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vd_empty_subfolders.R
\name{vd_empty_subfolders}
\alias{vd_empty_subfolders}
\title{Validate Which Folders Contain CSVs}
\usage{
vd_empty_subfolders(path, ...)
}
\arguments{
\item{path}{path to project directory}

\item{\dots}{ignored.}
}
\value{
Returns a list of validation results.
}
\description{
Validate Which Folders Contain CSVs
}
\examples{
dir_name <- file.path(tempdir(), "delete_me")
dir.create(dir_name)
dir(dir_name)
vd_empty_subfolders(dir_name)

sub_dirs <- file.path(dir_name, c('courses', 'students', 'faculty'))
lapply(sub_dirs, dir.create)
lapply(sub_dirs, dir.exists)
write.csv(mtcars, file=file.path(sub_dirs[3], 'mtcars.csv'))
dir(dir_name, recursive=TRUE, include.dirs =  TRUE, full.names = TRUE)
dir(dir_name, include.dirs =  TRUE, full.names = TRUE)
vd_empty_subfolders(dir(dir_name, include.dirs =  TRUE, full.names = TRUE))
}

