% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vd_multiple_csv.R
\name{vd_multiple_csv}
\alias{vd_multiple_csv}
\title{Validate Which Folders Contain Multiple CSVs}
\usage{
vd_multiple_csv(path, ...)
}
\arguments{
\item{path}{path to project directory}

\item{\dots}{ignored.}
}
\value{
Returns a list of validation results.
}
\description{
Validate Which Folders Contain Multiple CSVs
}
\examples{
dir_name <- file.path(tempdir(), "delete_me")
dir.create(dir_name)
dir(dir_name)
lapply(1:2, function(i) {
    write.csv(mtcars, file = file.path(dir_name, sprintf('file_\%s.csv', i)))
})
vd_multiple_csv(dir_name)
}
