% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vd_non_empty.R
\name{vd_non_empty}
\alias{vd_non_empty}
\title{Checks If Dir Not Empty}
\usage{
vd_non_empty(path, ...)
}
\arguments{
\item{path}{path to Dir}

\item{\dots}{ignored.}
}
\description{
Checks If Dir Not Empty
}
\examples{
dir.create("~/totally_rad_man")
	  vd_non_empty("~/totally_rad_man") 	#  FALSE
writeLines("hi", '~/filename1.csv')
	  vd_non_empty('~/filename1.csv') 	    #  TRUE
}

