% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vd_unmapped_csv.R
\name{vd_unmapped_csv}
\alias{vd_unmapped_csv}
\title{Validate Folders that Contain CSVs Are Found in Map}
\usage{
vd_unmapped_csv(paths, map, ...)
}
\arguments{
\item{paths}{paths to folders containing mapped csv files.}

\item{map}{A \pkg{valiData} map.}

\item{\dots}{ignored.}
}
\value{
Returns a list of validation results.
}
\description{
Validate Folders that contain CSVs are found in map
}

