% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vf_comma_broken.R
\name{vf_comma_broken}
\alias{comma_broken_report}
\alias{vf_comma_broken}
\title{Validate that a CSV's Is Properly Quoted}
\usage{
vf_comma_broken(path)

comma_broken_report(x, ...)
}
\arguments{
\item{path}{Path to a .csv file.}

\item{x}{A file or table validation function's (prefixed with \code{vf_} or
\code{vt_}) output.}

\item{\ldots}{ignored.}
}
\description{
\code{vf_comma_broken} - Validates that a .csv file has properly quoted character
columns (is not a broken .csv).

\code{comma_broken_report} - Generates accomanying report.
}

