% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vf_embedded_nul.R
\name{vf_embedded_nul}
\alias{vf_embedded_nul}
\title{Validate that a CSV's has No Embedded Nulls}
\usage{
vf_embedded_nul(path, ...)
}
\arguments{
\item{path}{Path to a .csv file.}

\item{\dots}{ignored.}
}
\description{
\code{vf_embedded_nul} - Validates that a .csv file has bno embedded nuls.  If
the file has embedded nuls it is likely due to not being in UTF-8 format (contains
multi-byte character sequences).
}
\examples{
loc <- file.path(tempdir(), 'temp.csv')
cat(
    paste(c(",x,y", "1, the dog, wen,1", "2,door,2"), collapse="\\n"),
    file=loc
)
vf_embedded_nul(loc)
}
