% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vf_file_type.R
\name{vf_file_type}
\alias{vf_file_type}
\title{Checks If File CSV}
\usage{
vf_file_type(file, type = "csv")
}
\arguments{
\item{file}{path to file}

\item{type}{extension of file representing type}
}
\description{
Checks If File CSV
}
\examples{
vf_file_type("broken_csv.R")           #  FALSE
vf_file_type("broken_csv.R", type="R") #  TRUE
vf_file_type("hello.csv")              #  TRUE
}

