% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{view_missing}
\alias{view_missing}
\title{view_missing}
\usage{
view_missing(df, ..., view = TRUE)
}
\arguments{
\item{df}{dataframe}

\item{...}{tidyselect}

\item{view}{logical. if false, returns tibble}
}
\value{
tibble
}
\description{
View rows of the dataframe where columns in the tidyselect specification contain missings
by default, detects missings in any column. The result is by default displayed in the viewer pane.
Can be returned as a tibble optionally.
}
\examples{

iris \%>\%
framecleaner::make_na(Species, vec = "setosa") \%>\%
view_missing(view = FALSE)
}
