% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_colnames.R
\name{vt_colnames}
\alias{colnames_report}
\alias{vt_colnames}
\title{Validate that a CSV's Columns Are Named Correctly}
\usage{
vt_colnames(data, map, ignore.case = FALSE, ignore.space = FALSE,
  file.name = NULL)

colnames_report(x, ...)
}
\arguments{
\item{data}{\code{\link[base]{data.frame}}.}

\item{map}{A \code{data.frame} with a \code{header} (header name) &
\code{required} (logical,; Is column required?).}

\item{ignore.case}{logical.  Should case be ignored?}

\item{ignore.space}{logical.  Should whitespace be ignored?}

\item{file.name}{An optional file name for use in reporting.}

\item{x}{A file or table validation function's (prefixed with \code{vf_} or
\code{vt_}) output.}

\item{\ldots}{ignored.}
}
\value{
Returns a list of validation results.
}
\description{
\code{vt_colnames} - Validates that a .csv file's
\code{\link[base]{data.frame}} columns are named correctly.

\code{colnames_report} - Generates accomanying report.
}
\examples{
set.seed(10)
map <- data.frame(
    header = c("foobar", colnames(mtcars)[-1]),
    required = sample(c(TRUE, FALSE), ncol(mtcars), TRUE), stringsAsFactors = FALSE
)

vt_colnames(mtcars, map)
colnames_report(vt_colnames(mtcars, map))
}

