% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vt_colorder.R
\name{vt_colorder}
\alias{colorder_report}
\alias{vt_colorder}
\title{Validate that a CSV's Columns Are Correctly Ordered}
\usage{
vt_colorder(data, map, file.name = NULL)

colorder_report(x, ...)
}
\arguments{
\item{data}{\code{\link[base]{data.frame}}.}

\item{map}{A \code{data.frame} with a \code{header} (header name) &
\code{required} (logical,; Is column required?).}

\item{file.name}{An optional file name for use in reporting.}

\item{x}{A file or table validation function's (prefixed with \code{vf_} or
\code{vt_}) output.}

\item{missing}{Values considered to be missing or null.}

\item{\ldots}{ignored.}
}
\value{
Returns a list of validation results.
}
\description{
\code{vt_colorder} - Validates that a .csv file's
\code{\link[base]{data.frame}} columns are ordered as expected.

\code{colorder_report} - Generates accomanying report.
}
\examples{
set.seed(10)
map <- data.frame(
    header = sample(colnames(mtcars)),
    required = sample(c(TRUE, FALSE), ncol(mtcars), TRUE), stringsAsFactors = FALSE
)

vt_colorder(mtcars, map)
colorder_report(vt_colorder(mtcars, map))
}

