% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_column_names.R
\name{vt_column_names}
\alias{vt_column_names}
\title{Validate that a CSV's Columns Are Named Correctly}
\usage{
vt_column_names(data, map, ignore.case = FALSE, ignore.space = FALSE,
  file_name = NULL)
}
\arguments{
\item{data}{\code{\link[base]{data.frame}}.}

\item{map}{A \code{data.frame} with a \code{header} (header name) &
\code{required} (logical,; Is column required?).}

\item{ignore.case}{logical.  Should case be ignored?}

\item{ignore.space}{logical.  Should whitespace be ignored?}

\item{file_name}{An optional file name for use in reporting.}
}
\value{
Returns a list of validation results.
}
\description{
\code{vt_column_names} - Validates that a .csv file's
\code{\link[base]{data.frame}} columns are named correctly.
}

