% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_column_order.R
\name{vt_column_order}
\alias{vt_column_order}
\title{Validate that a CSV's Columns Are Correctly Ordered}
\usage{
vt_column_order(data, map, file_name = NULL)
}
\arguments{
\item{data}{\code{\link[base]{data.frame}}.}

\item{map}{A \code{data.frame} with a \code{header} (header name) &
\code{required} (logical,; Is column required?).}

\item{file_name}{An optional file name for use in reporting.}
}
\value{
Returns a list of validation results.
}
\description{
\code{vt_column_order} - Validates that a .csv file's
\code{\link[base]{data.frame}} columns are ordered as expected.
}

