% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vf_comma_broken.R
\name{comma_broken_report}
\alias{comma_broken_report}
\title{Validate that a CSV's Is Properly Quoted}
\usage{
comma_broken_report(x, ...)
}
\arguments{
\item{x}{A file or table validation function's (prefixed with \code{vf_} or
\code{vt_}) output.}

\item{\ldots}{ignored.}
}
\description{
\code{comma_broken_report} - Generates accomanying report.
}

