% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_duplicated_rows.R
\name{vt_duplicated_rows}
\alias{vt_duplicated_rows}
\title{Validate a CSV Contains No Duplicate Rows}
\usage{
vt_duplicated_rows(data, file_name = NULL)
}
\arguments{
\item{data}{\code{\link[base]{data.frame}}.}

\item{file_name}{An optional file name for use in reporting.}
}
\value{
Returns a list of validation results.
}
\description{
\code{vt_duplicated_rows} - Validates that a .csv file's
\code{\link[base]{data.frame}} contains no duplicated rows.
}
\examples{
vt_duplicated_rows(CO2)
vt_duplicated_rows(CO2[, 1:3])
str(vt_duplicated_rows(CO2[, 1:3]))
}

