% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_enough_columns.R
\name{vt_enough_columns}
\alias{vt_enough_columns}
\title{Validate that a CSV's has N Columns}
\usage{
vt_enough_columns(data, map, file_name = NULL)
}
\arguments{
\item{data}{\code{\link[base]{data.frame}}.}

\item{map}{A \code{data.frame} with a \code{header} (header name) &
\code{required} (logical,; Is column required?).}

\item{file_name}{An optional file name for use in reporting.}
}
\value{
Returns a list of validation results.
}
\description{
\code{vt_enough_columns} - Validates that a .csv file's
\code{\link[base]{data.frame}} has n columns.
}

