% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_header.R
\name{vt_header}
\alias{header_report}
\alias{vt_header}
\title{Validate that a CSV Has a Header}
\usage{
vt_header(data, map, file.name = NULL)

header_report(x, ...)
}
\arguments{
\item{data}{\code{\link[base]{data.frame}}.}

\item{map}{A \code{data.frame} with a \code{header} (header name) &
\code{required} (logical,; Is column required?).}

\item{file.name}{An optional file name for use in reporting.}

\item{x}{A file or table validation function's (prefixed with \code{vf_} or
\code{vt_}) output.}

\item{\ldots}{ignored.}
}
\value{
Returns a list of validation results.
}
\description{
\code{vt_header} - Validates that a .csv file's
\code{\link[base]{data.frame}} contains no missing/null values in required
fields.

\code{header_report} - Generates accomanying report.
}
\examples{
set.seed(10)
map <- data.frame(
    header = colnames(mtcars),
    required = sample(c(TRUE, FALSE), ncol(mtcars), TRUE), stringsAsFactors = FALSE
)

df <- mtcars; colnames(df) <- mtcars[1, ]
vt_header(df, map)
header_report(vt_header(df, map))
}

