% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_na_rows.R
\name{vt_na_rows}
\alias{vt_na_rows}
\title{Find Rows with only NAs, If Any}
\usage{
vt_na_rows(data, file_name = NULL)
}
\arguments{
\item{data}{the data}

\item{file_name}{the file name}
}
\description{
Find Rows with only NAs, If Any.
}
\examples{
set.seed(10)
test <- data.frame(matrix(rep(NA,ncol(mtcars)*2), nrow=2))
test2 <- rbind(mtcars, stats::setNames( test, names(mtcars) ))

vt_na_rows(test2)
str(vt_na_rows(test2))
}

