% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_ncols.R
\name{vt_ncols}
\alias{ncols_report}
\alias{vt_ncols}
\title{Validate that a CSV's has N Columns}
\usage{
vt_ncols(data, map, file.name = NULL)

ncols_report(x, ...)
}
\arguments{
\item{data}{\code{\link[base]{data.frame}}.}

\item{map}{A \code{data.frame} with a \code{header} (header name) &
\code{required} (logical,; Is column required?).}

\item{file.name}{An optional file name for use in reporting.}

\item{x}{A file or table validation function's (prefixed with \code{vf_} or
\code{vt_}) output.}

\item{\ldots}{ignored.}
}
\value{
Returns a list of validation results.
}
\description{
\code{vt_ncols} - Validates that a .csv file's
\code{\link[base]{data.frame}} has n columns.

\code{ncols_report} - Generates accomanying report.
}
\examples{
set.seed(10)
map <- data.frame(
    header = colnames(mtcars)[-1],
    required = sample(c(TRUE, FALSE), -1 + ncol(mtcars), TRUE), stringsAsFactors = FALSE
)

vt_ncols(mtcars, map)
ncols_report(vt_ncols(mtcars, map))
}

