% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_non_ASCII.R
\name{vt_non_ASCII}
\alias{vt_non_ASCII}
\title{Validates and Reports If Non-ASCII characters present}
\usage{
vt_non_ASCII(data, ...)
}
\arguments{
\item{data}{the data to be tested}
}
\description{
Validates and Reports If Non-ASCII characters present
}

