% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_non_empty.R
\name{vt_non_empty}
\alias{vt_non_empty}
\title{Validate a CSV Contains Rows}
\usage{
vt_non_empty(data, file_name = NULL, ...)
}
\arguments{
\item{data}{\code{\link[base]{data.frame}}.}

\item{file_name}{An optional file name for use in reporting.}

\item{\ldots}{ignored.}
}
\value{
Returns a list of validation results.
}
\description{
\code{vt_non_empty} - Validates that a .csv file's
\code{\link[base]{data.frame}} contains rows.
}
\examples{
vt_non_empty(CO2)
vt_non_empty(CO2[0, 1:3])
str(vt_non_empty(CO2[0, 1:3]))
}

