% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vt_required.R
\name{vt_required}
\alias{required_report}
\alias{vt_required}
\title{Validate that a CSV Has No Missing/Null Values in Required Fields}
\usage{
vt_required(data, map, missing = c("", "NULL", "NA", "N/A", "na", "n/a"),
  file.name = NULL)

required_report(ls, ...)
}
\arguments{
\item{data}{\code{\link[base]{data.frame}}.}

\item{map}{A \code{data.frame} with a \code{header} (header name) &
\code{required} (logical,; Is column required?).}

\item{missing}{Values considered to be missing or null.}

\item{file.name}{An optional file name for use in reporting.}

\item{x}{A file or table validation function's (prefixed with \code{vf_} or
\code{vt_}) output.}

\item{\ldots}{ignored.}
}
\value{
Returns a list of validation results.
}
\description{
\code{vt_required} - Validates that a .csv file's
\code{\link[base]{data.frame}} contains no missing/null values in required
fields.

\code{required_report} - Generates accomanying report.
}
\examples{
set.seed(10)
map <- data.frame(
    header = colnames(mtcars),
    required = sample(c(TRUE, FALSE), ncol(mtcars), TRUE), stringsAsFactors = FALSE
)

df <- mtcars
df[c(1, 4), c("drat", "cyl")] <- NA
vt_required(df, map)
required_report(vt_required(df, map))
}

