% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_required_columns.R
\name{vt_required_columns}
\alias{vt_required_columns}
\title{Validate that a CSV Has No Missing/Null Values in Required Fields}
\usage{
vt_required_columns(data, map, missing = c("", "NULL", "NA", "N/A", "na",
  "n/a"), prop.acceptable = 0, file_name = NULL)
}
\arguments{
\item{data}{\code{\link[base]{data.frame}}.}

\item{map}{A \code{data.frame} with a \code{header} (header name) &
\code{required} (logical,; Is column required?).}

\item{missing}{Values considered to be missing or null.}

\item{prop.acceptable}{proportion of allowable missing}

\item{file_name}{An optional file name for use in reporting.}
}
\value{
Returns a list of validation results.
}
\description{
\code{vt_required_columns} - Validates that a .csv file's
\code{\link[base]{data.frame}} contains no missing/null values in required
fields.
}

