% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_required_headers.R
\name{vt_required_headers}
\alias{vt_required_headers}
\title{Validate that a CSV's Required Headers Are There}
\usage{
vt_required_headers(data, map, ignore.case = FALSE, ignore.space = FALSE,
  file_name = NULL)
}
\arguments{
\item{data}{\code{\link[base]{data.frame}}.}

\item{map}{A \code{data.frame} with a \code{header} (header name) &
\code{required} (logical,; Is column required?).}

\item{ignore.case}{logical.  Should case be ignored?}

\item{ignore.space}{logical.  Should whitespace be ignored?}

\item{file_name}{An optional file name for use in reporting.}
}
\value{
Returns a list of validation results.
}
\description{
\code{vt_required_headers} - Validates that a .csv file's
\code{\link[base]{data.frame}} required headers are there (differes from
\code{vt_required_columns} in that no values in cells are required just the
header).
}
