% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_spaced_colnames.R
\name{vt_spaced_colnames}
\alias{vt_spaced_colnames}
\title{Validate that a CSV's Column Names Contain No Spaces}
\usage{
vt_spaced_colnames(data, file_name = NULL)
}
\arguments{
\item{data}{\code{\link[base]{data.frame}}.}

\item{file_name}{An optional file name for use in reporting.}
}
\value{
Returns a list of validation results.
}
\description{
\code{vt_spaced_colnames} - Validates that a .csv file's
\code{\link[base]{data.frame}} column names contain no spaces.
}
\examples{
df <- mtcars; colnames(df)[c(1, 5)]  <- c("split header", "foo bar"); df
vt_spaced_colnames(df)
str(vt_spaced_colnames(df))
}

