% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm_overlap.R
\name{confirm_overlap}
\alias{confirm_overlap}
\title{Confirm Overlap}
\usage{
confirm_overlap(vec1, vec2, return_tibble = F)
}
\arguments{
\item{vec1}{vector 1}

\item{vec2}{vector 2}

\item{return_tibble}{logical. If TRUE, returns  a tibble. otherwise by default returns the database invisibly to be queried by helper functions.}
}
\value{
tibble. overlap summary or overlap table
}
\description{
A venn style summary of the overlap in unique values of 2 vectors
}
\examples{
confirm_overlap(iris$Sepal.Width, iris$Sepal.Length)
}
