% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataValidation-package.R
\docType{package}
\name{dataValidation-package}
\alias{dataValidation}
\alias{dataValidation-package}
\title{dataValidation: Validation For Data Frames}
\description{
Functions for validating the structure and properties of data frames.
}
\author{
\strong{Maintainer}: Harrison Tietze \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
