% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_distinct.R
\name{determine_distinct}
\alias{determine_distinct}
\title{Automatically determine primary key}
\usage{
determine_distinct(.data)
}
\arguments{
\item{.data}{a data frame}
}
\value{
none
}
\description{
Uses \code{confirm_distinct} in an iterative fashion to determine the primary key.
}
\examples{
iris \%>\% determine_distinct
}
