% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_mapping.R
\name{determine_mapping}
\alias{determine_mapping}
\title{Determine pairwise structural mappings}
\usage{
determine_mapping(df, ..., listviewer = TRUE)
}
\arguments{
\item{df}{a data frame}

\item{...}{columns or a tidyselect specification}

\item{listviewer}{logical. defaults to TRUE to view output using the listviewer package}
}
\value{
description of mappings
}
\description{
Determine pairwise structural mappings
}
\examples{

iris \%>\%
determine_mapping(listviewer = FALSE)
}
