% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose}
\alias{diagnose}
\title{diagnose}
\usage{
diagnose(df, ...)
}
\arguments{
\item{df}{dataframe}

\item{...}{tidyselect}
}
\value{
dataframe summary
}
\description{
this function is lifted from the excellent \code{dlookr} package. I have reproduced it here
because I consider it an essential validation function, but the \code{dlookr} package itself has
some dependencies which can be problematic for some machines to download.
}
\examples{
iris \%>\% diagnose()
}
