% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose}
\alias{diagnose}
\title{diagnose}
\usage{
diagnose(df, ...)
}
\arguments{
\item{df}{dataframe}

\item{...}{tidyselect}
}
\value{
dataframe summary
}
\description{
this function is inspired by the excellent \code{dlookr} package. It takes a dataframe and returns
a summary of unique and missing values of the columns.
}
\examples{
iris \%>\% diagnose()
}
