% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose_missing}
\alias{diagnose_missing}
\title{diagnose_missing}
\usage{
diagnose_missing(df, ...)
}
\arguments{
\item{df}{dataframe}

\item{...}{optional tidyselect}
}
\value{
tibble summary
}
\description{
faster than diagnose if emphasis is on diagnosing missing values. Also, only shows the columns with
any missing values.
}
\examples{

iris \%>\%
framecleaner::make_na(Species, vec = "setosa") \%>\%
diagnose_missing()
}
