% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_file.R
\name{print.validate_file}
\alias{print.validate_file}
\title{Prints a validate_file Object}
\usage{
\method{print}{validate_file}(x, ...)
}
\arguments{
\item{x}{A validate_file object.}

\item{as.report}{logical.  If \code{TRUE} a report will be generated in the
directory indicated by \code{x$path}.}

\item{delete}{logical.  If \code{TRUE} and \code{as.report = TRUE} any prior
instances of `report will be deleted.}

\item{\ldots}{ignored.}
}
\description{
Prints a validate_file object
}

