% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataValidation-package.R
\docType{package}
\name{valiData-package}
\alias{valiData}
\alias{valiData-package}
\title{valiData: valiData}
\description{
Functions for validating the structure and properties of data frames.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://harrison4192.github.io/valiData}
  \item \url{https://github.com/Harrison4192/valiData}
  \item Report bugs at \url{https://github.com/Harrison4192/valiData/issues}
}

}
\author{
\strong{Maintainer}: Harrison Tietze \email{Harrison4192@gmail.com}

}
\keyword{internal}
