% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valiData.R
\name{valiData}
\alias{valiData}
\title{Validate a Directory of Subdirectories of CSV Files}
\usage{
valiData(path, map, ...)
}
\arguments{
\item{path}{Path to a directory with drecories of .csv files.}

\item{map}{mapping of dictionary of data tests to be applied to import files.}

\item{\ldots}{ignored.}
}
\value{
Returns a valiData object.
}
\description{
Validate a directory of dubdirectories of .csv files.
}
\seealso{
\code{\link[valiData]{validate_file}}
}

