% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_file.R
\name{validate_file}
\alias{validate_file}
\title{Validate a CSV File}
\usage{
validate_file(file, core_data_map, column_map, ...)
}
\arguments{
\item{file}{A path to a .csv file to be validated.}

\item{map}{A \code{data.frame} with a \code{header} (header name) &
\code{required} (logical,; Is column required?).}

\item{\ldots}{ignored.}
}
\value{
Prints a report to the console of tests for file type, comma broken
csv, number of columns, spaces in column headers, correct column names,
required columns containing values.
}
\description{
Validate a .csv file.
}
\examples{
\dontrun{
validate_file("C:/Users/trinker/Desktop/myfolder/totest.csv", core_data_map[["totest"]])
}
}

