% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vc_categories.R
\name{vc_categories}
\alias{vc_categories}
\title{Validates and Reports If Correct Categories Used}
\usage{
vc_categories(x, levels = "the levels", colnames = "the column")
}
\arguments{
\item{x}{character vector}

\item{levels}{levels of the category}

\item{colnames}{vector's colname}
}
\description{
Validates and Reports If Correct Categories Used
}

