% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_cipcode.R
\name{vc_cipcode}
\alias{vc_cipcode}
\title{Validates If Valid CIP Code}
\usage{
vc_cipcode(data, x, ...)
}
\arguments{
\item{data}{A data frame.}

\item{x}{Column name from \code{data} (character string).}

\item{\dots}{ignored.}
}
\value{
Returns a \code{vc} classed list object.
}
\description{
Validates If Valid CIP Code
}
\examples{
dat <- data.frame(
    cips = c('9.99.99', '22', '22.', '22.34', '22.234', '22.2344', '22.2345', NA),
    stringsAsFactors = FALSE
)
vc_cipcode(dat, 'cips')
}

