% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_compare.R
\name{vc_compare}
\alias{vc_compare}
\title{Validates If Comparison Met}
\usage{
vc_compare(data, x, y, comparison, date = FALSE, ...)
}
\arguments{
\item{data}{A data frame.}

\item{x}{Column name from \code{data} (character string).}

\item{y}{Column name from \code{data} (character string) to be compared to.}

\item{comparison}{logical operator for the comparison}

\item{date}{logical.  If \code{TRUE} x and y are converted to dates via
\code{parsedate::parse_iso_8601}.}

\item{\ldots}{ignored.}
}
\description{
Validates If Comparison Met
}
\examples{
dat <- data.frame(
    a = c(NA, 1:10),
    b = c(0, 1, 2:10+1),
    d1 = c(NA, sprintf("2016-01-\%sT09", 11:20)),
    d2 = c("2016-01-01T09", "2016-01-11T09", sprintf("2016-05-\%sT09", 12:20)),
    stringsAsFactors = FALSE
)

vc_compare(dat, x = 'a', y = 'b', '<')
vc_compare(dat, x = 'a', y = 'b', '<=')
vc_compare(dat, x = 'd1', y = 'd2', '<', date=TRUE)
vc_compare(dat, x = 'd1', y = 'd2', '<=', date=TRUE)
}

