% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_compare.R
\name{vc_compare}
\alias{vc_compare}
\title{Validates and Reports If Comparison Met}
\usage{
vc_compare(x, y, comparison, colname_x = "the X column",
  colname_y = "the Y column", date = FALSE, ...)
}
\arguments{
\item{x}{character vector to be coerced}

\item{y}{character vector to be coerced and compared}

\item{comparison}{logical operator for the comparison}

\item{colname_x}{X vector's colname}

\item{colname_y}{Y vector's colname}

\item{date}{logical.  If \code{TRUE} x and y are converted to dates via
\code{parsedate::parse_iso_8601}.}

\item{\ldots}{ignored.}
}
\description{
Validates and Reports If Comparison Met
}

