% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_credits.R
\name{vc_credits}
\alias{vc_credits}
\title{Validates and Reports If Formatted Like Credit}
\usage{
vc_credits(x, colname_x = "the column")
}
\arguments{
\item{x}{character vector}

\item{colname_x}{vector's colname}
}
\description{
Validates and Reports If Formatted Like Credit
}

