% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_credits.R
\name{vc_credits}
\alias{vc_credits}
\title{Validates and If Formatted Like Credit}
\usage{
vc_credits(data, x, ...)
}
\arguments{
\item{data}{A data frame.}

\item{x}{Column name from \code{data} (character string).}

\item{\dots}{ignored.}
}
\description{
Validates and If Formatted Like Credit
}
\examples{
dat <- data.frame(
    credits = c(NA, 1.21, 1.34, '4.0', '5.0', .1, 12, 1111),
    stringsAsFactors = FALSE
)
vc_credits(dat, 'credits')
}

