% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_email.R
\name{vc_email}
\alias{vc_email}
\title{Validates If Email}
\usage{
vc_email(data, x, ...)
}
\arguments{
\item{data}{A data frame.}

\item{x}{Column name from \code{data} (character string).}

\item{\dots}{ignored.}
}
\description{
Validates If Email
}
\examples{
dat <- data.frame(
    email =c('cookie@cookiemonster.com', 'joe@hometown.uni'
    , 'meet@seven', '@Jim', 'joe@gmail.com', NA),
    stringsAsFactors = FALSE
)
vc_email(dat, 'email')
}

