% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vc_email.R
\name{vc_email}
\alias{vc_email}
\title{Validates and Reports If Email}
\usage{
vc_email(x, colnames = "the column")
}
\arguments{
\item{x}{character vector}

\item{colnames}{vector's colname}
}
\description{
Validates and Reports If Email
}

