% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vc_integer.R
\name{vc_integer}
\alias{vc_integer}
\title{Validates and Reports If Integer}
\usage{
vc_integer(x, colname_x = "the column")
}
\arguments{
\item{x}{character vector}

\item{colname_x}{vector's colname}
}
\description{
Validates and Reports If Integer
}

