% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vc_iso_datetime.R
\name{vc_iso_datetime}
\alias{vc_iso_datetime}
\title{Validates and Reports If ISO 8601}
\usage{
vc_iso_datetime(x, colnames = "the column")
}
\arguments{
\item{x}{character vector}

\item{colnames}{vector's colname}
}
\description{
Validates and Reports If ISO 8601
}

