% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_iso_datetime.R
\name{vc_iso_datetime}
\alias{vc_iso_datetime}
\title{Validates and Reports If ISO 8601}
\usage{
vc_iso_datetime(x, colname_x = "the column")
}
\arguments{
\item{x}{character vector}

\item{colname_x}{vector's colname}
}
\description{
Validates and Reports If ISO 8601
}

