% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vc_logical.R
\name{vc_logicial}
\alias{vc_logicial}
\title{Validates and Reports If Logical}
\usage{
vc_logicial(x, colnames = "the column")
}
\arguments{
\item{x}{character vector}

\item{colnames}{vector's colname}
}
\description{
Validates and Reports If Logical
}

