% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_missing.R
\name{vc_missing}
\alias{vc_missing}
\title{Checks For Missing Elements In Vector}
\usage{
vc_missing(x, prop.acceptable = 0, missing = c("", "NULL", "NA", "N/A",
  "na", "n/a"), required = FALSE, file.name = NULL)
}
\arguments{
\item{x}{char vector}

\item{prop.acceptable}{proportion of allowable missing}

\item{missing}{values for missing can take}

\item{required}{whether the vector in the table is required}

\item{file.name}{name of file containing the table of vectors}
}
\description{
Checks For Missing Elements In Vector
}
\examples{
set.seed(10)
x <- sample(unlist(c(missing, NA, rep(LETTERS[1:10], 100))), 10000, T)
vc_missing(x)
}

