% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_null.R
\name{vc_null}
\alias{vc_null}
\title{Checks For Null Elements In Vector}
\usage{
vc_null(x, prop.acceptable = 0, null = c("", "NULL"), required = FALSE,
  file.name = NULL)
}
\arguments{
\item{x}{char vector}

\item{prop.acceptable}{proportion of allowable null}

\item{required}{whether the vector in the table is required}

\item{file.name}{name of file containing the table of vectors}

\item{missing}{values null can take}
}
\description{
Checks For Null Elements In Vector
}
\examples{
set.seed(10)
x <- sample(unlist(c(missing, NA, rep(LETTERS[1:10], 100))), 10000, T)
vc_null(x)
}

