% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_numeric.R
\name{vc_numeric}
\alias{vc_numeric}
\title{Validates and Reports If Numeric}
\usage{
vc_numeric(x, colname_x = "the column")
}
\arguments{
\item{x}{character vector}

\item{colname_x}{vector's colname}
}
\description{
Validates and Reports If Numeric
}

