% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_numeric.R
\name{vc_numeric}
\alias{vc_numeric}
\title{Validates and Reports If Numeric}
\usage{
vc_numeric(data, x, ...)
}
\arguments{
\item{data}{A data frame.}

\item{x}{Column name from \code{data} (character string).}

\item{\dots}{ignored.}
}
\description{
Validates and Reports If Numeric
}
\examples{
dat <- data.frame(
    vals = c('9.99.99', '22', '22.', '22.34', '22.234', '23,456', '12,23', 'dog', NA),
    vals2 = c('123', '123,456', '-1,234,123.456', '-1,234,123', '.2', NA, 1, 0, ""),
    stringsAsFactors = FALSE
)
vc_numeric(dat, 'vals')
vc_numeric(dat, 'vals2')
}

