% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vc_regex.R
\name{vc_regex}
\alias{vc_regex}
\title{Validates and Reports If Formatted Like Regex Provided}
\usage{
vc_regex(x, regex, colname_x = "the column")
}
\arguments{
\item{x}{character vector}

\item{regex}{regex}

\item{colname_x}{vector's colname}
}
\description{
Validates and Reports If Formatted Like Regex Provided
}

