% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_type.R
\name{vc_type}
\alias{vc_type}
\title{Switch Method Based On Type Of Data in Column}
\usage{
vc_type(x, type, colname_x = "the column")
}
\arguments{
\item{x}{character vector}

\item{type}{:  date_utc, datetime, email, enumerated, integer, logical, numeric, zipcode, cipcode
guid,enumerated are handled outside by \code{vc_unique} and \code{vc_categories}}

\item{colname_x}{the vector's colname}
}
\description{
Switch Method Based On Type Of Data in Column
}

