% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vc_unique.R
\name{vc_unique}
\alias{vc_unique}
\title{Validates and Reports If Unique}
\usage{
vc_unique(x, colnames = "the column")
}
\arguments{
\item{x}{character vector}

\item{colnames}{vector's colname}
}
\description{
Validates and Reports If Unique
}

