% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_utc_date.R
\name{vc_utc_date}
\alias{vc_utc_date}
\title{Validates and Reports If Date UTC}
\usage{
vc_utc_date(x, colname_x = "the column")
}
\arguments{
\item{x}{character vector}

\item{colname_x}{vector's colname}
}
\description{
Validates and Reports If Date UTC
}

