% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_utc_date.R
\name{vc_utc_date}
\alias{vc_utc_date}
\title{Validates If Date UTC}
\usage{
vc_utc_date(data, x, ...)
}
\arguments{
\item{data}{A data frame.}

\item{x}{Column name from \code{data} (character string).}

\item{\dots}{ignored.}
}
\description{
Validates If Date UTC
}
\examples{
dat <- data.frame(
    start = c("2016-01-01T09", "R2D2-3CPO", sprintf("2016-04-\%sT09", 12:19), NA),
    end = c(NA, sprintf("2016-01-\%sT09", 11:20)),
    stringsAsFactors = FALSE
)

vc_utc_date(dat, 'start')
vc_utc_date(dat, 'end')
}

