% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_zipcode.R
\name{vc_zipcode}
\alias{vc_zipcode}
\title{Validates and Reports If ZipCode}
\usage{
vc_zipcode(x, colname_x = "the column")
}
\arguments{
\item{x}{character vector}

\item{colname_x}{vector's colname}
}
\description{
Validates and Reports If ZipCode
}

