% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vc_zipcode.R
\name{vc_zipcode}
\alias{vc_zipcode}
\title{Validates and Reports If ZipCode}
\usage{
vc_zipcode(x, colnames = "the column")
}
\arguments{
\item{x}{character vector}

\item{colnames}{vector's colname}
}
\description{
Validates and Reports If ZipCode
}

