% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vf_comma_broken.R
\name{vf_comma_broken}
\alias{vf_comma_broken}
\title{Validate that a CSV's Is Properly Quoted}
\usage{
vf_comma_broken(path)
}
\arguments{
\item{path}{Path to a .csv file.}
}
\description{
\code{vt_comma_broken} - Validates that a .csv file has properly quoted character
columns (is not a broken .csv).
}

