% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vf_csv_broken.R
\name{vf_csv_broken}
\alias{vf_csv_broken}
\title{Validate that a CSV's Is Properly Quoted}
\usage{
vf_csv_broken(path)
}
\arguments{
\item{path}{Path to a .csv file.}
}
\description{
\code{vt_comma_broken} - Validates that a .csv file has properly quoted character
columns (is not a broken .csv).
}

