% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vf_empty_subfolders.R, R/vf_file_type.R
\name{vf_empty_subfolders}
\alias{report_empty_subfolders}
\alias{report_file_type}
\alias{vf_empty_subfolders}
\alias{vf_file_type}
\title{Validate Which Folders Contain CSVs}
\usage{
vf_empty_subfolders(all_subpaths, csv_subpaths)

report_empty_subfolders(x, ...)

vf_file_type(file, type = "csv")

report_file_type(x, ...)
}
\arguments{
\item{all_subpaths}{absolute paths from root to all subfolders under path, see documentation for \code{get_paths_root_to_files(path)}}

\item{csv_subpaths}{paths to all subfolders CONTAINING CSVs, see documentation for \code{get_paths_to_csvs(path)}}

\item{x}{output of \code{vf_empty_subfolders()}}

\item{file}{path to file}

\item{type}{extension of file representing type}

\item{\ldots}{extra parameters}

\item{x}{output of \code{vf_file_type()}}

\item{\ldots}{extra parameters}
}
\value{
Returns a list of validation results.
}
\description{
Validate Which Folders Contain CSVs

Checks If File CSV
}
\examples{
x <- vf_file_type("broken_csv.R")           #  FALSE
x <- vf_file_type("broken_csv.R", type="R") #  TRUE
x <- vf_file_type("hello.csv")              #  TRUE
report_file_type(x)
}

