% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vd_empty_subfolders.R
\name{vd_empty_subfolders}
\alias{report_empty_subfolders}
\alias{vd_empty_subfolders}
\title{Validate Which Folders Contain CSVs}
\usage{
vd_empty_subfolders(all_subpaths, csv_subpaths)

report_empty_subfolders(x, ...)
}
\arguments{
\item{all_subpaths}{absolute paths from root to all subfolders under path, see documentation for \code{get_paths_root_to_files(path)}}

\item{csv_subpaths}{paths to all subfolders CONTAINING CSVs, see documentation for \code{get_paths_to_csvs(path)}}

\item{x}{output of \code{vd_empty_subfolders()}}

\item{\ldots}{extra parameters}
}
\value{
Returns a list of validation results.
}
\description{
Validate Which Folders Contain CSVs
}

