% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vf_file_type.R
\name{vf_file_type}
\alias{report_file_type}
\alias{vf_file_type}
\title{Checks If File CSV}
\usage{
vf_file_type(file, type = "csv")

report_file_type(x, ...)
}
\arguments{
\item{file}{path to file}

\item{type}{extension of file representing type}

\item{x}{output of \code{vf_file_type()}}

\item{\ldots}{extra parameters}
}
\description{
Checks If File CSV
}
\examples{
x <- vf_file_type("broken_csv.R")           #  FALSE
x <- vf_file_type("broken_csv.R", type="R") #  TRUE
x <- vf_file_type("hello.csv")              #  TRUE
report_file_type(x)
}

