% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_duplicated_rows.R
\name{vt_duplicated_rows}
\alias{duplicated_rows_report}
\alias{vt_duplicated_rows}
\title{Validate a CSV Contains No Duplicate Rows}
\usage{
vt_duplicated_rows(data, file.name = NULL)

duplicated_rows_report(x, ...)
}
\arguments{
\item{data}{\code{\link[base]{data.frame}}.}

\item{file.name}{An optional file name for use in reporting.}

\item{x}{A file or table validation function's (prefixed with \code{vf_} or
\code{vt_}) output.}

\item{\ldots}{ignored.}
}
\value{
Returns a list of validation results.
}
\description{
\code{vt_duplicated_rows} - Validates that a .csv file's
\code{\link[base]{data.frame}} contains no duplicated rows.

\code{report_duplicated_rows} - Generates accomanying report.
}
\examples{
vt_duplicated_rows(CO2)
vt_duplicated_rows(CO2[, 1:3])
duplicated_rows_report(vt_duplicated_rows(CO2[, 1:3]))
}

