% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_header.R
\name{vt_header}
\alias{vt_header}
\title{Validate that a CSV Has a Header}
\usage{
vt_header(data, map, file_name = NULL)
}
\arguments{
\item{data}{\code{\link[base]{data.frame}}.}

\item{map}{A \code{data.frame} with a \code{header} (header name) &
\code{required} (logical,; Is column required?).}

\item{file_name}{An optional file name for use in reporting.}
}
\value{
Returns a list of validation results.
}
\description{
\code{vt_header} - Validates that a .csv file's
\code{\link[base]{data.frame}} contains no missing/null values in required
fields.
}
\examples{
set.seed(10)
map <- data.frame(
    header = colnames(mtcars),
    required = sample(c(TRUE, FALSE), ncol(mtcars), TRUE), stringsAsFactors = FALSE
)

df <- mtcars; colnames(df) <- mtcars[1, ]
vt_header(df, map)
str(vt_header(df, map))
}

