% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vt_non_ASCII.R
\name{vt_non_ASCII}
\alias{vt_non_ASCII}
\title{Validates and Reports If Non-ASCII characters present}
\usage{
vt_non_ASCII(data, ...)
}
\arguments{
\item{data}{the data to be tested}

\item{\ldots}{ignored}
}
\description{
Validates and Reports If Non-ASCII characters present
}
\examples{
dat <- data.frame(
    txt = c(
        "fa\\xE7ile test of showNonASCII():",
        "\\\\details{",
        "   This is a good line",
        "   This has an \\xfcmlaut in it.",
        "   OK again.",
        "}"
    ),
    x = 1:6,
    stringsAsFactors = FALSE
)
vt_non_ASCII(dat)
str(vt_non_ASCII(dat))
}

