% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vt_non_empty.R
\name{vt_non_empty}
\alias{non_empty_report}
\alias{vt_non_empty}
\title{Validate a CSV Contains Rows}
\usage{
vt_non_empty(data, file.name = NULL)

non_empty_report(x, ...)
}
\arguments{
\item{data}{\code{\link[base]{data.frame}}.}

\item{file.name}{An optional file name for use in reporting.}

\item{x}{A file or table validation function's (prefixed with \code{vf_} or
\code{vt_}) output.}

\item{\ldots}{ignored.}
}
\value{
Returns a list of validation results.
}
\description{
\code{vt_non_empty} - Validates that a .csv file's
\code{\link[base]{data.frame}} contains rows.

\code{report_duplicated_rows} - Generates accomanying report.
}
\examples{
vt_non_empty(CO2)
vt_non_empty(CO2[, 1:3])
duplicated_non_empty(vt_non_empty(CO2[, 1:3]))
}

