% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vt_spaced_colnames.R
\name{vt_spaced_colnames}
\alias{spaced_colnames_report}
\alias{vt_spaced_colnames}
\title{Validate that a CSV's Column Names Contain No Spaces}
\usage{
vt_spaced_colnames(data, file.name = NULL)

spaced_colnames_report(x, ...)
}
\arguments{
\item{data}{\code{\link[base]{data.frame}}.}

\item{file.name}{An optional file name for use in reporting.}

\item{x}{A file or table validation function's (prefixed with \code{vf_} or
\code{vt_}) output.}

\item{\ldots}{ignored.}
}
\value{
Returns a list of validation results.
}
\description{
\code{vt_spaced_colnames} - Validates that a .csv file's
\code{\link[base]{data.frame}} column names contain no spaces.

\code{spaced_colnames_report} - Generates accomanying report.
}
\examples{
df <- mtcars; colnames(df)[c(1, 5)]  <- c("split header", "foo bar"); df
vt_spaced_colnames(df)
spaced_colnames_report(vt_spaced_colnames(df))
}

